backend_flask/requirements.txt
Flask==3.1.0
torch==2.6.0
torchvision==0.21.0
easyocr==1.7.2
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
scikit-learn==1.6.1
Werkzeug==3.1.3
python-dotenv
openai
langdetect